+++ src/data/about.ts
@@ -14,7 +14,7 @@ const about: IAboutItem[] = [
   },
   {
     icon: {
-      color: colors.pink,
+      color: colors.primary,
       name: "code-fork",
       type: "font-awesome"
     },
@@ -23,7 +23,7 @@ const about: IAboutItem[] = [
   },
   {
     icon: {
-      color: colors.orange,
+      color: colors.primary,
       name: "gitlab",
       type: "font-awesome"
     },
@@ -32,17 +32,17 @@ const about: IAboutItem[] = [
   },
   {
     icon: {
-      color: colors.green,
+      color: colors.primary,
       name: "web",
       type: "material-community"
     },
     title: "Personal Website",
-    url: "https://hmajid2301.github.io"
+    url: "https://haseebmajid.dev"
   },
   {
     function_to_call: "store",
     icon: {
-      color: colors.red,
+      color: colors.primary,
       name: "rate-review",
       type: "material"
     },